Remove AnimationProvider context, derive frames from Date.now()
Replace the AnimationProvider React context with a simpler approach:
- Animation frames are computed from Date.now() in useStatusIcon
- A single setInterval in DevUI drives re-renders when actively
  building/syncing, and stops when idle/error/synced
- Eliminates dev-ui-animation-context.tsx entirely
