+++ packages/twenty-sdk/src/cli/utilities/dev/ui/components/dev-ui.tsx
@@ -2,12 +2,12 @@ import { type OrchestratorStateEvent } from '@/cli/utilities/dev/orchestrator/de
 import { DevUiApplicationPanel } from '@/cli/utilities/dev/ui/components/dev-ui-application-panel';
 import { DevUiEntityLegend } from '@/cli/utilities/dev/ui/components/dev-ui-entity-section';
 import { DevUiEventItem } from '@/cli/utilities/dev/ui/components/dev-ui-event-log';
-import { AnimationProvider } from '@/cli/utilities/dev/ui/dev-ui-animation-context';
 import { InkProvider, useInk } from '@/cli/utilities/dev/ui/dev-ui-ink-context';
 import { type DevUiStateManager } from '@/cli/utilities/dev/ui/dev-ui-state-manager';
 import React, { useReducer, useEffect } from 'react';
 
 const ACTIVE_PIPELINE_STATUSES = new Set(['building', 'syncing']);
+const ANIMATION_TICK_MS = 120;
 
 const DevUI = ({
   uiStateManager,
@@ -23,10 +23,20 @@ const DevUI = ({
   }, [uiStateManager]);
 
   const state = uiStateManager.getSnapshot();
-  const isPaused = !ACTIVE_PIPELINE_STATUSES.has(state.pipeline.status);
+  const isActive = ACTIVE_PIPELINE_STATUSES.has(state.pipeline.status);
+
+  useEffect(() => {
+    if (!isActive) {
+      return;
+    }
+
+    const timer = setInterval(() => forceRender(), ANIMATION_TICK_MS);
+
+    return () => clearInterval(timer);
+  }, [isActive]);
 
   return (
-    <AnimationProvider paused={isPaused}>
+    <>
       <Static items={state.events}>
         {(event: OrchestratorStateEvent) => (
           <DevUiEventItem key={event.id} event={event} />
@@ -37,7 +47,7 @@ const DevUI = ({
         <DevUiApplicationPanel state={state} />
         <DevUiEntityLegend />
       </Box>
-    </AnimationProvider>
+    </>
   );
 };
